modified:   src/controllers/auth.controller.js 	modified:   src/controllers/carts.controller.js 	modified:   src/controllers/mail.controller.js 	modified:   src/controllers/user.controller.js 	modified:   src/model/DAOs/products/products.mongo.dao.js 	modified:   src/model/DAOs/tickets/tickets.mongo.doa.js 	modified:   src/model/DAOs/users/users.dao.js 	new file:   src/public/documents/account/6525c8fbd1769faa3ea88267_1696976586421-Untitled-1-1.png 	new file:   src/public/documents/adress/6525c8fbd1769faa3ea88267_1696976586467-Untitled-1-1.png 	new file:   src/public/documents/info/6525c8fbd1769faa3ea88267_1696976586551-Untitled-1-1.png 	modified:   src/public/js/views.js 	new file:   src/public/store/1696977882066-Untitled-1-1.png 	modified:   src/routes/profile.router.js 	modified:   src/services/auth.service.js 	modified:   src/services/carts.service.js 	new file:   src/services/mail.service.js 	modified:   src/services/product.service.js 	modified:   src/services/ticket.service.js 	modified:   src/views/profile.handlebars 	modified:   src/views/ticketsView.handlebars

+++ src/model/DAOs/users/users.dao.js
@@ -11,7 +11,10 @@ class UsersDao {
     }
 
     async findByIdAndDelete(id, index) {
-        return await userModel.findByIdAndDelete(id[index])
+        if (index) {
+            return await userModel.findByIdAndDelete(id[index])
+        }
+        return await userModel.findByIdAndDelete(id)
     }
 
     async findByIdLean(id) {
@@ -52,6 +55,9 @@ class UsersDao {
         return await userModel.findOne(data)
     }
 
+    async findLean() {
+        return await userModel.find().lean()
+    }
 }
 
 module.exports = UsersDao
